test: line/section test 분리
